refactor: move things to separate classes
