+++ codegen/src/generators/ResourcesGenerator/ResourceFactory.ts
@@ -1,0 +1,8 @@
+/**
+ * Creates a resource interface.
+ * @example
+ * export interface PetsResource {
+ *  getPetById(params: GetPetByIdParams): Promise<GetPetByIdResponse>
+ * }
+ */
+export class ResourceFactory {}
