+++ codegen/src/typescript/ClassBuilder.ts
@@ -14,7 +14,7 @@ export class ClassBuilder {
     private readonly parameters: Property[] = []
     private readonly properties: Property[] = []
 
-    public constructor(private readonly className: string, private readonly context: CodeGenContext) { }
+    public constructor(private readonly className: string, private readonly context: CodeGenContext) {}
 
     public addParameter(param: Property) {
         this.parameters.push(param)
@@ -35,22 +35,22 @@ export class ClassBuilder {
         return this.context.prettifyTS(`
             export class ${this.className}${this.extend ? ` extends ${this.extend.className}` : ''} {
                 ${this.properties
-                .map(
-                    prop =>
-                        `public readonly ${prop.name}${prop.initialValue ? '' : '?'}: ${prop.type}${
-                        prop.initialValue ? ` = ${prop.initialValue}` : ''
-                        }`
-                )
-                .join('\n')}
+                    .map(
+                        prop =>
+                            `public readonly ${prop.name}${prop.initialValue ? '' : '?'}: ${prop.type}${
+                                prop.initialValue ? ` = ${prop.initialValue}` : ''
+                            }`
+                    )
+                    .join('\n')}
                 public constructor(
                     ${this.parameters
-                .map(
-                    param =>
-                        `public readonly ${param.name}: ${param.type}${
-                        param.initialValue ? ` = ${param.initialValue}` : ''
-                        }`
-                )
-                .join(',\n')}
+                        .map(
+                            param =>
+                                `public readonly ${param.name}: ${param.type}${
+                                    param.initialValue ? ` = ${param.initialValue}` : ''
+                                }`
+                        )
+                        .join(',\n')}
                 ) { ${this.extend ? this.extend.constructorCalls.join('\r') : ''}}
             }
         `)
